holiday productivity update
code optimization and organization, fade ins/outs, pause sound filter

+++ Assets/Environment/lives/LifeUI.cs
@@ -5,7 +5,6 @@ using UnityEngine;
 public class LifeUI : MonoBehaviour
 {
     internal Animator animator => GetComponent<Animator>();
-    internal int health => (int)GameObject.FindWithTag("Player").GetComponent<PlayerControl>().health;
-
+    internal int health => GameObject.FindWithTag("Player").GetComponent<PlayerControl>().health;
     void Update() => animator.SetInteger("vida", health);
 }
